+++ R/flextable_sizes.R
@@ -553,11 +553,14 @@ text_metric <- function(x) {
   txt_data
 }
 
+#' @importFrom grDevices dev.cur dev.set
 strings_sizes_tmp <- function(...) {
+  old_dev <- dev.cur()
   file <- tempfile(fileext = ".png")
   agg_png(width = 50, height = 50, filename = file, units = "in")
   on.exit({
     dev.off()
+    if (old_dev > 1L) dev.set(old_dev)
     unlink(file, force = TRUE)
   })
   strings_sizes(...)