+++ R/rtf_str.R
@@ -213,9 +213,3 @@ rtf_add.flextable <- function(x, value, ...) {
   x$content[[length(x$content) + 1]] <- value
   x
 }
-
-
-#' @importFrom officer to_rtf to_html
-.onLoad <- function(libname, pkgname) {
-  # registerS3method("to_rtf", "flextable", to_rtf.flextable)
-}